# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IndiLingo — Indian regional language learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /languages/{languageId}/units:
    get:
      operationId: getLanguageUnits
      tags: [languages]
      summary: Get units for a language (no user progress)
      parameters:
        - in: path
          name: languageId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Units without user progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitWithProgress"
        "404":
          description: Language not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/languages/{languageId}/units:
    get:
      operationId: getUserLanguageUnits
      tags: [users]
      summary: Get units for a language with user progress
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: path
          name: languageId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Units with user progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitWithProgress"
        "404":
          description: Language or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson with its exercises
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson with exercises
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonWithExercises"
        "404":
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [lessons]
      summary: Mark a lesson as complete (server-authoritative XP/streak/stars)
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonCompleteInput"
      responses:
        "200":
          description: Completion result with updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonCompleteResult"
        "404":
          description: Lesson or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LessonSummary:
      type: object
      required: [id, unitId, title, orderIndex, xpReward]
      properties:
        id:
          type: string
        unitId:
          type: string
        title:
          type: string
        orderIndex:
          type: integer
        xpReward:
          type: integer
        stars:
          type: ["integer", "null"]
        completed:
          type: boolean
    UnitWithProgress:
      type: object
      required: [id, languageId, title, description, orderIndex, unitType, lessons, completedLessons, totalLessons]
      properties:
        id:
          type: string
        languageId:
          type: string
        title:
          type: string
        description:
          type: string
        orderIndex:
          type: integer
        unitType:
          type: string
          enum: [script, vocabulary, phrases, conversation]
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonSummary"
        completedLessons:
          type: integer
        totalLessons:
          type: integer
    Exercise:
      type: object
      required: [id, lessonId, type, question, correctAnswer, options]
      properties:
        id:
          type: string
        lessonId:
          type: string
        type:
          type: string
          enum: [script_practice, multiple_choice, translate, fill_blank, match_pairs]
        question:
          type: string
        correctAnswer:
          type: string
        options:
          type: array
          items:
            type: string
        nativeScript:
          type: ["string", "null"]
        romanization:
          type: ["string", "null"]
    LessonWithExercises:
      type: object
      required: [id, unitId, title, orderIndex, xpReward, exercises]
      properties:
        id:
          type: string
        unitId:
          type: string
        title:
          type: string
        orderIndex:
          type: integer
        xpReward:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
    LessonCompleteInput:
      type: object
      required: [userId, correctCount]
      properties:
        userId:
          type: string
        correctCount:
          type: integer
    LessonCompleteResult:
      type: object
      required: [xpEarned, stars, streakDays, totalXp]
      properties:
        xpEarned:
          type: integer
        stars:
          type: integer
        streakDays:
          type: integer
        totalXp:
          type: integer
...