# lib/api-zod/src/generated/types/lessonCompleteResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IndiLingo — Indian regional language learning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LessonCompleteResult {
  xpEarned: number;
  stars: number;
  streakDays: number;
  totalXp: number;
}